+++ src/core/component/directives/in-view/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.0.0-rc.60 (2020-09-01)
+
+#### :house: Internal
+
+* [Split the module into two: API was moved to `core/dom/in-view`](https://github.com/V4Fire/Client/issues/310)
+
 ## v3.0.0-rc.51 (2020-08-04)
 
 #### :bug: Bug Fix